- fix delete image cache "node allready deleted do not use"
git-svn-id: https://server.mycore.de/svn/mycore/trunk@14512 7089d063-bc14-496d-ac90-53de1872eb6c

+++ components/imaging/sources/org/mycore/services/imaging/MCRImgCacheEventHandler.java
@@ -49,7 +49,7 @@ public class MCRImgCacheEventHandler extends MCREventHandlerBase {
      */
     protected void handleFileDeleted(MCREvent evt, MCRFile file) {
         MCRImgCacheManager imgCache = MCRImgCacheManager.instance();
-        if (imgCache.existInCache(file)) {
+        if (!file.getOwnerID().equals(MCRImgCacheManager.CACHE_FOLDER) && imgCache.existInCache(file)) {
             try {
                 MCRImgCacheCommands.deleteCachedFile(file.getID());
             } catch (Exception e) {
